Function ic_crypto::crypto_hash

pub fn crypto_hash<T: CryptoHashable>(data: &T) -> CryptoHashOf<T>

Creates a (typed) domain-separated cryptographic hash.

The bytes that are hashed are a combination of

the byte representation of the hash domain obtained via CryptoHashables supertrait CryptoHashDomain

the bytes fed to the hasher state via CryptoHashables supertrait Hash

Note that the trait CryptoHashDomain is sealed for security reasons. To implement this trait for a new struct that shall be cryptographically hashed, contact the crypto team.

The (secure) hashing algorithm that is used internally is intentionally unspecified because it may be subject to change across registry/protocol versions. Use Sha256 instead if the algorithm used for producing the hash must not change across registry/protocol versions.