Struct ic_crypto::CryptoComponentFatClient

pub struct CryptoComponentFatClient<C: CryptoServiceProvider> { /* fields omitted */ }

Allows Internet Computer nodes to perform crypto operations such as distributed key generation, signing, signature verification, and TLS handshakes.

Implementations

impl<R: Rng + CryptoRng + Send + Sync> CryptoComponentFatClient<Csp<R, ProtoSecretKeyStore>>

pub fn new_with_rng_and_fake_node_id(
csprng: R,
config: &CryptoConfig,
logger: ReplicaLogger,
registry_client: Arc<dyn RegistryClient>,
node_id: NodeId
) -> Self

Creates a crypto component using the given csprng and fake node_id.

impl<C: CryptoServiceProvider> CryptoComponentFatClient<C>

pub fn new_with_csp_and_fake_node_id(
csp: C,
logger: ReplicaLogger,
registry_client: Arc<dyn RegistryClient>,
node_id: NodeId
) -> Self

Creates a crypto component using the given csp and fake node_id.

impl CryptoComponentFatClient<Csp<OsRng, ProtoSecretKeyStore>>

pub fn new(
config: &CryptoConfig,
registry_client: Arc<dyn RegistryClient>,
logger: ReplicaLogger,
metrics_registry: Option<&MetricsRegistry>
) -> Self

Creates a new crypto component.

This is the constructor to use to create the replica's / node's crypto component.

Multiple crypto components must share the same state to avoid problems due to concurrent state access. To achieve this, we recommend to instantiate multiple components as in the example below.

WARNING: Multiple crypto componets must be instantiated with Arc::clone as in the example. Do not create multiple crypto componets with the same config (as opposed to using Arc::clone), as this will lead to concurrency issues e.g. when the components access the secret key store simultaneously.

use ic_config::crypto::CryptoConfig;
use ic_crypto::CryptoComponent;
use ic_logger::replica_logger::no_op_logger;
use std::sync::Arc;
use ic_registry_client::fake::FakeRegistryClient;
use ic_registry_common::proto_registry_data_provider::ProtoRegistryDataProvider;
use ic_crypto::utils::get_node_keys_or_generate_if_missing;
use ic_metrics::MetricsRegistry;

CryptoConfig::run_with_temp_config(|config| {
    // instantiate a registry somehow
    let registry_client = FakeRegistryClient::new(Arc::new(ProtoRegistryDataProvider::new()));

    // get a logger and metrics registry
    let logger = no_op_logger();
    let metrics_registry = MetricsRegistry::new();

    let first_crypto_component = Arc::new(CryptoComponent::new(&config, Arc::new(registry_client), logger, Some(&metrics_registry)));
    let second_crypto_component = Arc::clone(&first_crypto_component);
});

pub fn new_with_fake_node_id(
config: &CryptoConfig,
registry_client: Arc<dyn RegistryClient>,
node_id: NodeId,
logger: ReplicaLogger
) -> Self

Creates a crypto component using a fake node_id.

pub fn new_for_non_replica_process(
config: &CryptoConfig,
registry_client: Arc<dyn RegistryClient>,
logger: ReplicaLogger
) -> impl CryptoComponentForNonReplicaProcess

Creates a crypto component that offers limited functionality and can be used outside of the replica process.

Please refer to the trait documentation of CryptoComponentForNonReplicaProcess for more details.

pub fn get_node_id(&self) -> NodeId

Returns the NodeId of this crypto component.