Trait ic_crypto::CryptoComponentForNonReplicaProcess

pub trait CryptoComponentForNonReplicaProcess: KeyManager + ThresholdSigVerifierByPublicKey<CatchUpContentProtobufBytes> + TlsHandshake + Send + Sync { }

A crypto component that offers limited functionality and can be used outside of the replica process.

This is an intermediate solution before crypto runs in a separate process.

This should be used whenever crypto is required on a node, but a full-fledged CryptoComponent is not available. Example use cases are in separate process such as ic-fe or the node manager.

Do not instantiate a CryptoComponent outside of the replica process, since that may lead to problems with concurrent access to the secret key store. CryptoComponentForNonReplicaProcess guarantees that only methods are exposed that don't risk running into such concurrency issues, as they do not modify the secret key store.