Function ic_crypto::utils::generate_initial_dkg_encryption_keys

pub fn generate_initial_dkg_encryption_keys(
    crypto_root: &Path, 
    dkg_config: &InitialDkgConfig, 
    node_id: NodeId
) -> CryptoResult<EncryptionPublicKeyWithPop>

Generates initial DKG ephemeral encryption keys for the dkg_config and the node_id.

The secret key is stored in the crypto secret key store at the crypto_root.

The corresponding public key is returned

If the crypto root does not exist yet, it is created and the appropriate permissions are set.