Function ic_crypto::utils::get_node_keys_or_generate_if_missing_for_node_id

pub fn get_node_keys_or_generate_if_missing_for_node_id(
    crypto_root: &Path, 
    node_id: NodeId
) -> NodePublicKeys

Obtains the node keys or generates them for the given node_id if they are missing.

Retrieves the node's public keys from crypto_root, and checks their consistency with the secret keys. If there are no public keys in crypto_root, new keys are generated: the secret parts are stored in the secret key store and the public parts are stored in the public key store.

Panics

if public keys exist but are inconsistent with the secret keys.

if an error occurs when accessing or generating the key.

Notes

There are two variants of the function:

get_node_keys_or_generate_if_missing_for_node_id() which takes an additional node_id argument, and uses it when generating TLS keys. This is a temporary variant intended for the setup when node ids are generated externally.

get_node_keys_or_generate_if_missing() which derives node_id from the retrieved or generated node signing public key, uses the derived node_id when generating TLS keys, and returns node_id together with all the generated public keys. This is the variant to be used eventually by node manager and other clients.